Add Community Algorithm
Add Community Algorithm: this uses the Louvain method to find the
modularity of the nodes in the graph, then finds an appropriate
resolution to break the graph into partitions. A selectedNode with the
highest betweenness is found for each partition. An edge is created
between the newcomer and the selectedNodes (starting with the
selectedNode from the largest partition).